Add MRI tests for Logger.
Now that the licensing is square, just copy this stuff straight out of
stdlib.

We need to skip the tests that test log rotation, as we can't actually
rotate.

In order to get the tests running, it was found that the signature to
create a logger changed, so that fixes a bug as well.

Finally, we need to be locked to MiniTest 4 because the tests were
written under it, and I don't want to touch them since they're
essentially vendored.